+++ sparse_dot_mkl/__init__.py
@@ -1,11 +1,22 @@
-__version__ = '0.9.2'
+__version__ = '0.9.3'
 
 
 from sparse_dot_mkl.sparse_dot import (
     dot_product_mkl,
     dot_product_transpose_mkl,
-    get_version_string,
     gram_matrix_mkl,
     sparse_qr_solve_mkl,
     set_debug_mode
 )
+
+from ._mkl_interface import (
+    mkl_get_max_threads,
+    mkl_set_interface_layer,
+    mkl_set_num_threads,
+    mkl_get_version,
+    mkl_get_version_string,
+    mkl_set_num_threads_local,
+    mkl_interface_integer_dtype
+)
+
+get_version_string = mkl_get_version_string